% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{check_p_value}
\alias{check_p_value}
\title{return confidence indicator[. * ** ***] according to the value X}
\usage{
check_p_value(x)
}
\arguments{
\item{x}{numeric}
}
\description{
return confidence indicator[. * ** ***] according to the value X
}
