% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BL.R
\name{coef.ridgereg}
\alias{coef.ridgereg}
\title{B coefficients}
\usage{
\method{coef}{ridgereg}(object, ...)
}
\arguments{
\item{object}{Refers to the object return by ridge regression}

\item{...}{argument to be passed to methods}
}
\value{
Coefficients
}
\description{
B coefficients
}
\examples{
data(iris)
s<-ridgereg(formula=Petal.Length~Species, data=iris, 2)
coef(s)
}
