% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{get_png}
\alias{get_png}
\title{get PNG file(used in plot())
private function}
\usage{
get_png(filename)
}
\arguments{
\item{filename}{png file name used in plot}
}
\description{
get PNG file(used in plot())
private function
}
