\name{Lab04-package}
\alias{Lab04-package}
\alias{Lab04}
\docType{package}
\title{
\packageTitle{Lab04}
}
\description{
\packageDescription{Lab04}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lab04}
\packageIndices{Lab04}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lab04}

Maintainer: \packageMaintainer{Lab04}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
