% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\docType{data}
\name{linkoping_theme}
\alias{linkoping_theme}
\title{customize Liu theme (used in plot())
private theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 4.
}
\usage{
linkoping_theme
}
\description{
customize Liu theme (used in plot())
private theme
}
\keyword{datasets}
