% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{This function uses RC class to generate Linear Regressions}
\value{
This class returns linear regression data including t-values and p-values
}
\description{
This function uses RC class to generate Linear Regressions
}
\section{Fields}{

\describe{
\item{\code{formula}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{reg_coef}}{matrix.}

\item{\code{fitted_val}}{matrix.}

\item{\code{residuals}}{matrix.}

\item{\code{degree_of_freedom}}{numeric.}

\item{\code{residul_variance}}{numeric.}

\item{\code{variance_reg_coef}}{matrix.}

\item{\code{t_value}}{matrix.}

\item{\code{p_value}}{matrix.}
}}


\examples{
\dontrun{
data("iris")
item <- linreg(formula = Petal.Length~Species, data = iris)
}
}
