% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{linreg.coef}
\alias{linreg.coef}
\title{Coefficients}
\usage{
linreg.coef(obj)
}
\arguments{
\item{obj}{}
}
\value{
obj$coefficients
}
\description{
Coefficients
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          linreg.coef(s)
}
