% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Linear Regression Model using QR decomposition}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{A model formula equation}

\item{data}{A data set}
}
\value{
Multiple regression coefficients
}
\description{
It compute multiple regression model values to determine statistical parameters.
}
\examples{
lm<- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
}
