% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{linreg.resid}
\alias{linreg.resid}
\title{Residual of the linear regression function}
\usage{
linreg.resid(obj)
}
\arguments{
\item{obj}{}
}
\value{
obj$residuals
}
\description{
Residual of the linear regression function
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          linreg.print(s)
}
