% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{linreg.summary}
\alias{linreg.summary}
\title{Summary of linear regression function}
\usage{
linreg.summary(obj)
}
\arguments{
\item{obj}{}
}
\value{
data frame with the final computations
}
\description{
Summary of linear regression function
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          linreg.summary(s)
}
