% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BL.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(error)
}
\arguments{
\item{error}{resiudals errors in model}
}
\value{
mae_res Mean absolute error value
}
\description{
Mean Absolute Error
}
\examples{
data(iris)
ridgeres <- ridgereg(formula= Sepal.Length ~ Petal.Length + Petal.Width,
data = iris, lambda= 2)
x= data.frame(Petal.Length= iris$Petal.Length, Petal.Width = iris$Petal.Width)
rr_error <- iris$epal.Length-predict(ridgeres,x)
mae_ridge <- mae(rr_error)
}
