% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Model diagnostic plots}
\usage{
\method{plot}{linreg}(obj)
}
\arguments{
\item{obj}{object of type linreg}
}
\value{
ggplot object
}
\description{
Model diagnostic plots
}
\examples{
data(iris)
s <- linreg(formula=Sepal.Length ~ Petal.Length + Petal.Width, data=iris)
plot(s)
}
