% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{pred}
\alias{pred}
\title{Predicted values}
\usage{
pred(obj)
}
\arguments{
\item{obj}{Refers to the object return by the linear regression function}
}
\value{
obj$fitted_values
}
\description{
Returns the predicted values
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          pred(s)
}
