% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BL.R
\name{predict.ridgereg}
\alias{predict.ridgereg}
\title{Predict}
\usage{
\method{predict}{ridgereg}(object, ndf, ...)
}
\arguments{
\item{object}{Refers to the object return by ridge regression}

\item{ndf}{New data frame to predict}

\item{...}{argument to be passed to methods}
}
\value{
fitted values
}
\description{
Predict
}
\examples{
data(iris)
          s<-ridgereg(formula=Petal.Length~Species, data=iris, 2)
          predict(s)
}
