% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{print.linreg}
\alias{print.linreg}
\title{Methods of function linreg}
\usage{
\method{print}{linreg}(obj)
}
\arguments{
\item{obj}{Refers to the res variable return by the linear regression function}
}
\value{
c, obj$residuals, obj$fitted_values, format(obj$coefficients),
}
\description{
Methods of function linreg
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data=iris)
          print(s)

}
