% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print}
\alias{print}
\title{print}
\usage{
print(results)
}
\arguments{
\item{results}{Argument}

\item{...}{Additional Argument}
}
\description{
print
}
\examples{
mod <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
print(mod)
}
