% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BL.R
\name{print.ridgereg}
\alias{print.ridgereg}
\title{Print}
\usage{
\method{print}{ridgereg}(x, ...)
}
\arguments{
\item{x}{Refers to the object return by ridge regression}

\item{...}{argument to be passed to methods}
}
\value{
Coefficients
}
\description{
Print model call coefficients
}
\examples{
data(iris)
          s<-ridgereg(formula=Petal.Length~Species, data=iris, 2)
          print(s)
}
