% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qr_hr.R
\name{qr_hr}
\alias{qr_hr}
\title{QR factorization with Householder reflection}
\usage{
qr_hr(A)
}
\arguments{
\item{A}{n x p (p <= n) matrix for which to do the QR factorization}
}
\value{
list with Q and R matrices
}
\description{
QR factorization with Householder reflection
}
\examples{
data(iris)
formula <- Sepal.Length ~ Petal.Length + Petal.Width
data <- iris
X <-stats::model.matrix(object=formula, data=data)
QR <- qr_hr(X)
}
