% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{resid.linreg}
\alias{resid.linreg}
\title{Residual of the linear regression function}
\usage{
\method{resid}{linreg}(obj)
}
\arguments{
\item{obj}{Refers to the res variable return by the linear regression function}
}
\value{
obj$residuals
}
\description{
Returns the residuals
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          resid(s)
}
