% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{resid}
\alias{resid}
\title{Residual()}
\usage{
resid(results, ...)
}
\arguments{
\item{results}{Arguments}

\item{...}{Additional Arguments}
}
\description{
Residual()
}
\examples{
mod <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
resid(mod)
}
