% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resid.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{This function is returning the Residuals form linreg}
\usage{
\method{residuals}{linreg}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
This function is returning the Residuals form linreg
}
\examples{
lm<-linreg(formula = Petal.Length ~ Species, data = iris)
resid(lm)
}
