% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ridgereg.R
\name{ridgereg}
\alias{ridgereg}
\title{Ridge regression}
\usage{
ridgereg(formula, data, lambda = 0)
}
\arguments{
\item{formula}{The Y and X variables of the regression}

\item{data}{A data frame}

\item{lambda}{Hyperparameter of the regression}
}
\value{
res
}
\description{
Ridge regression
}
\examples{
data(iris)
          ridgereg(Petal.Length~Species, iris,2)
}
