% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_BL.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(error)
}
\arguments{
\item{error}{residuals errors in model}
}
\value{
rmse_res root mean squared error value
}
\description{
Root Mean Squared Error
}
\examples{
data(iris)
ridgeres <- ridgereg(formula= Sepal.Length ~ Petal.Length + Petal.Width,
data = iris, lambda= 2)
x= data.frame(Petal.Length= iris$Petal.Length, Petal.Width = iris$Petal.Width)
rr_error <- iris$epal.Length-predict(ridgeres,x)
rmse_ridge <- rmse(rr_error)
}
