% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{summary()}
\usage{
summary(results, ...)
}
\arguments{
\item{results}{Argument}

\item{...}{Additional Argument}
}
\value{
Return the summary of linreq
}
\description{
summary()
}
\examples{
mod <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
summary(mod)
}
