% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{coef.linreg}
\alias{coef.linreg}
\title{Coefficients}
\usage{
\method{coef}{linreg}(obj)
}
\arguments{
\item{obj}{Refers to the res variable return by the linear regression function}
}
\value{
obj$coefficients
}
\description{
Coefficients
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          coef(s)
}
