% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.linreg}
\alias{coef.linreg}
\title{This function is returning the regression coefficients form linreg}
\usage{
\method{coef}{linreg}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
This function is returning the regression coefficients form linreg
}
\examples{
lm<-linreg(formula = Petal.Length ~ Species, data = iris)
coef(lm)
}
