% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04-package.R
\docType{package}
\name{lab04-package}
\alias{lab04-package}
\alias{lab04}
\title{Lab 4 of Advanced Programming in R}
\description{
This package include following S3 class\cr
\cr 
1. linreg - a class provide linear regression  that is similar as lm()\cr 
\cr
}
