% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{linreg}
\description{
This class perform linear regression
}
\section{Fields}{

\describe{
\item{\code{data_name}}{character.}

\item{\code{form}}{formula.}

\item{\code{data}}{data.frame.}

\item{\code{Regressions_coefficients}}{vector.}

\item{\code{The_fitted_value}}{matrix.}

\item{\code{The_residuals}}{matrix.}

\item{\code{The_degrees_of_freedom}}{numeric.}

\item{\code{The_variance_of_the_reg_coef}}{matrix.}

\item{\code{t_values}}{list.}
}}


