% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Multiple regression model}
\value{

}
\description{
Multiple regression model
}
\section{Fields}{

\describe{
\item{\code{Formula}}{formula.}

\item{\code{RegressionCoeficients}}{matrix.}

\item{\code{FittedValues}}{matrix.}

\item{\code{Residuals}}{matrix.}

\item{\code{DegreesOfFreedom}}{numeric.}

\item{\code{ResidualVariance}}{matrix.}

\item{\code{VarianceOfTheRegressionCoefficients}}{matrix.}

\item{\code{TValues}}{vector.}

\item{\code{DataName}}{character.}

\item{\code{Pvalues}}{vector.}
}}


