% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Multiple regression model}
\value{
class
}
\description{
Multiple regression model
}
\section{Fields}{

\describe{
\item{\code{Formula}}{formula.}

\item{\code{RegressionCoeficients}}{matrix.}

\item{\code{FittedValues}}{matrix.}

\item{\code{Residuals}}{matrix.}

\item{\code{DegreesOfFreedom}}{numeric.}

\item{\code{ResidualVariance}}{matrix.}

\item{\code{VarianceOfTheRegressionCoefficients}}{matrix.}

\item{\code{TValues}}{vector.}

\item{\code{DataName}}{character.}

\item{\code{Pvalues}}{vector.}
}}

\section{Methods}{

\describe{
\item{\code{coef()}}{returns coefficients as a named vector}

\item{\code{initialize(formula, data)}}{constructor}

\item{\code{plot()}}{plots residuals vs Fitted values plot and Scale - Location plot}

\item{\code{pred()}}{returns the predicted values}

\item{\code{print()}}{prints formula, name of data frame and regression coeficients}

\item{\code{resid()}}{returns vector of residuals}

\item{\code{summary()}}{returns the summary of linear regression model}
}}

