% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Lab4: Computations using ordinary least squares, implement methods in RC class}
\arguments{
\item{FunctionFormula}{formula object passed by user}

\item{formula}{formula object passed by the user}

\item{data}{dataframe passed by the user}

\item{y}{dependent variables matrix}

\item{RegCofMatrix}{independent variables matrix}

\item{X}{independent variables matrix}
}
\value{
class

creates an object based on the arguments passed
Setter for regression coefficient

the value that has been set
Setter for Fitted Values

the value that has been set
Setter for Fitted Values

the value that has been set
Setter for degree of freedom

the value that has been set
Setter for degree of residual variance

the value that has been set
Setter for degree of variance of reg coefficients

the value that has been set
Setter T Values

the value that has been set
Setter P Values

the value that has been set
Print like lm

nothing
Plot using ggplot2

nothing
Residuals calculation

Residuals
predicted values

FittedValues
coefficeints as named vec

coefficeints as named vec
Print the Summary

coefficeints as named vec
}
\description{
Constructor to initialize the data fields of this class

Sets the value for RegressionCoeficientMatrix field

Sets the value for FittedValues

Sets the value for Residuals

Sets the value for DegreesOfFreedom

Sets the value for ResidualVariance

Sets the value for VarianceOfTheRegressionCoefficients

Sets the value for TValues

Sets the value for TValues

Prints out the coefficient and coefficent names

Plots 2 graphs on a grid that are mention in the lab manual

return only residuals as vec form

return the preticted values (y-hat)

return the coefficeints as named vec

Print the simmary along with the pvalue, tvalue, sigma and degree of freedom
}
\section{Fields}{

\describe{
\item{\code{VarianceOfTheRegressionCoefficients}}{matrix.}

\item{\code{RegressionCoeficientMatrix}}{matrix.}

\item{\code{TValues}}{t-values in vec form}

\item{\code{Residuals}}{matrix.}

\item{\code{ResidualVariance}}{matrix.}

\item{\code{DegreesOfFreedom}}{numeric.}

\item{\code{DataName}}{Name of the dataframe passed.}

\item{\code{FittedValues}}{matrix.}

\item{\code{PValues}}{t-values in vec form}
}}

\section{Methods}{

\describe{
\item{\code{print()}}{lllllllllllllllllllllllllllllllllllllllll}
}}

\examples{
newobj= linreg$new(frm, iris)
newobj$summary()

}
