% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\docType{class}
\name{linreg-class}
\alias{linreg-class}
\alias{linreg}
\title{Lab4: Computations using ordinary least squares, implement methods in RC class}
\arguments{
\item{FunctionFormula}{formula object passed by user}

\item{data}{dataframe passed by the user}

\item{RegCofMatrix}{independent variables matrix}

\item{X}{independent variables matrix}
}
\value{
class

the value that has been set
}
\description{
Sets the value for FittedValues
}
\section{Fields}{

\describe{
\item{\code{VarianceOfTheRegressionCoefficients}}{matrix.}

\item{\code{RegressionCoeficientMatrix}}{matrix.}

\item{\code{TValues}}{t-values in vec form}

\item{\code{Residuals}}{matrix.}

\item{\code{ResidualVariance}}{matrix.}

\item{\code{DegreesOfFreedom}}{numeric.}

\item{\code{DataName}}{Name of the dataframe passed.}

\item{\code{FittedValues}}{matrix.}

\item{\code{PValues}}{t-values in vec form}
}}

\section{Methods}{

\describe{
\item{\code{pred()}}{return the preticted values (y-hat)}

\item{\code{print()}}{Prints out the coefficient and coefficent names}

\item{\code{resid()}}{return only residuals as vec form}

\item{\code{setDegreesOfFreedom(RegCofMatrix, X)}}{Sets the value for DegreesOfFreedom}

\item{\code{setFittedValues(RegCofMatrix, X)}}{Sets the value for FittedValues}

\item{\code{setPValues()}}{Sets the value for TValues}

\item{\code{setRegCofficient(X, y)}}{Sets the value for RegressionCoeficientMatrix field}

\item{\code{setResidual(y)}}{Sets the value for Residuals}

\item{\code{setResidualVariance()}}{Sets the value for ResidualVariance}

\item{\code{setTValues()}}{Sets the value for TValues}

\item{\code{setVarianceOfTheRegressionCoefficients(X)}}{Sets the value for VarianceOfTheRegressionCoefficients}

\item{\code{summary()}}{Print the simmary along with the pvalue, tvalue, sigma and degree of freedom}
}}

