% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Multiple regression model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}
}
\value{
S3 object of linreg
}
\description{
Class takes linear equation and data structure, calculates regression coefficients, fitted values, residuals, degrees of freedom, residual variance, variance of the regression coefficients, t-values and p-values.
}
