% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linreg.R
\name{linreg}
\alias{linreg}
\title{Multiple regression model}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{formula}

\item{data}{data.frame}
}
\value{
S3 object of linreg
}
\description{
The algorithm takes a graph and an initial node and calculates the shortest path from the initial node to every other node in the graph. Find Wikipedia docs \href{https://cutt.ly/SfE5Er1}{Here}.
}
