% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{linreg}
\alias{linreg}
\title{A S3 class provide linear regression  that is similar as lm().}
\usage{
linreg(formula, data)
}
\arguments{
\item{formula}{Formula}

\item{data}{Dataframe}
}
\value{
linreg object(list) contain all the calculated values regarding the linear regression
}
\description{
A S3 class provide linear regression  that is similar as lm().
}
