% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{scale - location and Residuals vs Fitted values plots}
\usage{
\method{plot}{linreg}(obj)
}
\arguments{
\item{obj}{linreg object}
}
\value{
plots
}
\description{
takes linreg object and plots scale - location and Residuals vs
Fitted values plots.
}
