% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{This function is being used to plot regression coefficients values form linreg}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
This function is being used to plot regression coefficients values form linreg
}
\examples{
lm<-linreg(formula = Petal.Length ~ Species, data = iris)
plot(lm)

}
