% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.linreg}
\alias{plot.linreg}
\title{Plots for a linreg object}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments that we don't use}
}
\description{
Plots for a linreg object
}
\examples{
fit <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
plot(fit)
}
