% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{plot.linreg}
\alias{plot.linreg}
\title{plot 2 graphs with Liu Logo and customized liu theme.}
\usage{
\method{plot}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg objec}

\item{...}{graphical parameters to plot}
}
\description{
plot 2 graphs with Liu Logo and customized liu theme.
}
