% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Pred()}
\usage{
pred(results, ...)
}
\arguments{
\item{results}{Aurgument}

\item{...}{Additional Argument}
}
\value{

}
\description{
Pred()
}
\examples{
mod <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
pred(mod)
}
