% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{Get fitted values from a linreg object}
\usage{
pred(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments that we don't use}
}
\description{
Get fitted values from a linreg object
}
\examples{
fit <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
pred(fit)
}
