% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred.R
\name{pred}
\alias{pred}
\title{This function is returning the fitted values form linreg}
\usage{
pred(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
This function is returning the fitted values form linreg
}
\examples{
lm<-linreg(formula = Petal.Length ~ Species, data = iris)
pred(lm)
}
