% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{This function printing the regression coefficients form linreg}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object}

\item{...}{Additional arguments}
}
\description{
This function printing the regression coefficients form linreg
}
\examples{
lm<-linreg(formula = Petal.Length ~ Species, data = iris)
print(lm)
}
