% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.linreg}
\alias{print.linreg}
\title{Print for a linreg object}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{An object of class linreg}

\item{...}{Additional arguments that we don't use}
}
\description{
Print for a linreg object
}
\examples{
fit <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
print(fit)
}
