% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{print.linreg}
\alias{print.linreg}
\title{print out formula, data and Coefficients of current linear regression.}
\usage{
\method{print}{linreg}(x, ...)
}
\arguments{
\item{x}{linreg object}

\item{...}{other objects}
}
\description{
print out formula, data and Coefficients of current linear regression.
}
