% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{residuals.linreg}
\alias{residuals.linreg}
\title{vector of residuals}
\usage{
\method{residuals}{linreg}(obj)
}
\arguments{
\item{obj}{linreg object}
}
\value{
vector of residuals
}
\description{
vector of residuals
}
