% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods1.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Summary of linear regression function}
\usage{
\method{summary}{linreg}(obj)
}
\arguments{
\item{obj}{Refers to the res variable return by the linear regression function}
}
\value{
data frame with the final computations
}
\description{
Summary of linear regression function
}
\examples{
data(iris)
          s<-linreg(formula=Petal.Length~Species, data= iris)
          summary(s)
}
