% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.linreg}
\alias{summary.linreg}
\title{Printing the summary of linreg object}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Additional arguments}
}
\description{
Printing the summary of linreg object
}
\examples{
lm <- linreg(Petal.Length ~ Sepal.Width + Sepal.Length, data = iris)
summary(lm)
}
