% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab04.R
\name{summary.linreg}
\alias{summary.linreg}
\title{print out summary which is simiar as lm()}
\usage{
\method{summary}{linreg}(object, ...)
}
\arguments{
\item{object}{linreg object}

\item{...}{other objects}
}
\description{
print out summary which is simiar as lm()
}
